Make interpreter async (#130)

+++ air/src/execution_step/air/fold/fold_state.rs
@@ -32,7 +32,7 @@ pub(crate) struct FoldState<'i> {
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub(crate) enum IterableType {
     Scalar,
-    Stream(usize),
+    Stream(u32),
 }
 
 impl<'i> FoldState<'i> {
